+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix deprecated warning on DocLinkBox component.
+
 ## [2.85.1] - 2020-08-14
 
 Fix: